+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/compile/Validations.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.engine.compile
 import cats.data.Validated.valid
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{MissingParameters, RedundantParameters}
 import pl.touk.nussknacker.engine.api.context._
-import pl.touk.nussknacker.engine.api.definition.{Parameter, ParameterValidator}
+import pl.touk.nussknacker.engine.api.definition.{Parameter, Validator}
 import pl.touk.nussknacker.engine.api.expression.{TypedExpression, TypedExpressionMap}
 import pl.touk.nussknacker.engine.api.{NodeId, ParameterNaming}
 import pl.touk.nussknacker.engine.compiledgraph.TypedParameter
@@ -31,16 +31,17 @@ object Validations {
   }
 
   def validateWithCustomValidators(
-      parameters: List[(TypedParameter, Parameter)]
+      parameters: List[(TypedParameter, Parameter)],
+      paramValidatorsMap: Map[String, ValidatedNel[PartSubGraphCompilationError, List[Validator]]]
   )(
       implicit nodeId: NodeId
-  ): ValidatedNel[PartSubGraphCompilationError, List[(TypedParameter, Parameter)]] = {
-    val validationResults = for {
-      (typedParam, paramDef) <- parameters
-      paramValidationResult = validate(paramDef.validators, typedParam)
-    } yield paramValidationResult
-    validationResults.sequence.map(_ => parameters)
-  }
+  ): ValidatedNel[PartSubGraphCompilationError, List[(TypedParameter, Parameter)]] =
+    parameters
+      .map { case (typedParam, _) =>
+        paramValidatorsMap(typedParam.name).andThen(validator => validate(validator, typedParam))
+      }
+      .sequence
+      .map(_ => parameters)
 
   private def validateRedundancy(definedParamNamesSet: Set[String], usedParamNamesSet: Set[String])(
       implicit nodeId: NodeId
@@ -62,22 +63,23 @@ object Validations {
     validate(paramDefinition.validators, parameter._1).map((_, parameter._2))
   }
 
-  def validate(validators: List[ParameterValidator], parameter: TypedParameter)(
+  def validate(validators: List[Validator], parameter: TypedParameter)(
       implicit nodeId: NodeId
   ): ValidatedNel[PartSubGraphCompilationError, TypedParameter] = {
+    val paramWithValueAndExpressionList = parameter.typedValue match {
+      case te: TypedExpression => List((parameter.name, te.typingInfo.typingResult.valueOpt, te.expression))
+      case tem: TypedExpressionMap =>
+        tem.valueByKey.toList.map { case (branchName, expression) =>
+          (
+            ParameterNaming.getNameForBranchParameter(parameter.name, branchName),
+            expression.returnType.valueOpt,
+            expression.expression
+          )
+        }
+    }
+
     validators
       .flatMap { validator =>
-        val paramWithValueAndExpressionList = parameter.typedValue match {
-          case te: TypedExpression => List((parameter.name, te.typingInfo.typingResult.valueOpt, te.expression))
-          case tem: TypedExpressionMap =>
-            tem.valueByKey.toList.map { case (branchName, expression) =>
-              (
-                ParameterNaming.getNameForBranchParameter(parameter.name, branchName),
-                expression.returnType.valueOpt,
-                expression.expression
-              )
-            }
-        }
         paramWithValueAndExpressionList.map { case (name, value, expression) =>
           validator.isValid(name, Expression(expression.language, expression.original), value, None).toValidatedNel
         }
